Move and rename autoAdminOnRegister() cloud function
