+++ functions/src/utils/user.js
@@ -1,0 +1,17 @@
+const admin = require("firebase-admin");
+
+/** Add a user to the list of admins
+ *
+ * @param {string} newAdminId
+ */
+const addAdmin = async (newAdminId) => {
+  await admin
+    .firestore()
+    .collection("roles")
+    .doc("admin")
+    .update({
+      users: admin.firestore.FieldValue.arrayUnion(newAdminId),
+    });
+};
+
+exports.addAdmin = addAdmin;
