+++ functions/venue.js
@@ -3,6 +3,7 @@ const functions = require("firebase-functions");
 const { HttpsError } = require("firebase-functions/lib/providers/https");
 
 const { checkAuth } = require("./src/utils/assert");
+const { addAdmin } = require("./src/utils/user");
 const { getVenueId, checkIfValidVenueId } = require("./src/utils/venue");
 
 const PLAYA_VENUE_ID = "jamonline";
@@ -406,20 +407,6 @@ const dataOrUpdateKey = (data, updated, key) =>
     typeof updated[key] !== "undefined" &&
     updated[key]);
 
-/** Add a user to the list of admins
- *
- * @param {string} newAdminId
- */
-const addAdmin = async (newAdminId) => {
-  await admin
-    .firestore()
-    .collection("roles")
-    .doc("admin")
-    .update({
-      users: admin.firestore.FieldValue.arrayUnion(newAdminId),
-    });
-};
-
 /** Remove a user from the list of admins
  *
  * @param {string} adminId
@@ -433,13 +420,6 @@ const removeAdmin = async (adminId) => {
       users: admin.firestore.FieldValue.arrayRemove(adminId),
     });
 };
-
-exports.createUser = functions.auth.user().onCreate(async (user) => {
-  const flag = functions.config().flag;
-  if (flag && flag.autoadmin) {
-    await addAdmin(user.uid);
-  }
-});
 
 exports.addVenueOwner = functions.https.onCall(async (data, context) => {
   checkAuth(context);
